refactor : convertor 패키지 생성 및 메서드명 변경
Generator 클래스를 Convertor 라는 이름으로 변경하였으므로 convertor라는 패키지를 만들어 따로 분리했습니다.
메서드 명도 convertor로 시작하게 변경했습니다.
