+++ src/main/java/racingcar/controller/RaceGameController.java
@@ -5,8 +5,8 @@ import java.util.stream.Collectors;
 
 import racingcar.domain.Car;
 import racingcar.domain.Cars;
-import racingcar.util.generator.NameConvertor;
-import racingcar.util.generator.NumberConvertor;
+import racingcar.util.convertor.NameConvertor;
+import racingcar.util.convertor.NumberConvertor;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -30,7 +30,7 @@ public class RaceGameController {
 	}
 
 	private List<Car> addCars() {
-		List<String> carNames = NameConvertor.generateCarNames(InputView.getCarNames());
+		List<String> carNames = NameConvertor.convertCarNames(InputView.getCarNames());
 		return carNames.stream()
 			.map(Car::new)
 			.collect(Collectors.toList());
@@ -38,7 +38,7 @@ public class RaceGameController {
 
 	private int getAttemptNumber() {
 		try {
-			return NumberConvertor.generateNumber(InputView.getAttemptNumber());
+			return NumberConvertor.convertNumber(InputView.getAttemptNumber());
 		} catch (IllegalArgumentException e) {
 			OutputView.printErrorMessage(e.getMessage());
 			return getAttemptNumber();
