+++ src/main/java/racingcar/util/convertor/NumberConvertor.java
@@ -1,9 +1,9 @@
-package racingcar.util.generator;
+package racingcar.util.convertor;
 
 import racingcar.util.validator.NumberValidator;
 
 public class NumberConvertor {
-	public static int generateNumber(String attemptNumber) {
+	public static int convertNumber(String attemptNumber) {
 		NumberValidator.validateNumber(attemptNumber);
 		return Integer.parseInt(attemptNumber);
 	}
